broke up stack constraints into multiple files
